Added hasNext() and updated getNext().
Added new function hasNext for checking meta data in case it has the next request to make to get remaining data.  Store response internally so we don't have to pass anything to hasNext() or getNext() and it updates internal as it's called.  Also make it easier for getting the next result set if available for requests.